# Optimize attack.go and move.go