Changes in `internal/action/step/attack.go` file:

Cast Rate Improvements:
- Added FCR-adjusted attack cycle timing, implemented burst skill timeout handling, optimized skill selection latency (-40ms per cast)

Safety Enhancements:
- Added monster position validation, improved stuck detection algorithm, reduced melee repositioning threshold by 25%

Changes in `internal/action/step/move.go` file:
Anti-Idle Optimization:
- Reduced idle detection threshold from 3s => 1.5s
- Shortened anti-idle recovery sleep from 300ms => 150ms
- Added position validation before random movement

FCR Enhancement:
- Implemented cast duration multiplier (7/10 base value, line <126> for additional changes of FCR)
- Added safety checks for minimum sleep duration and unified teleport/vigor skill handling


NOT tested on all zones, though Countess/Cows/Ancient Tunnels/Lower Kurast Chests/CS/Baal seem to work well without errors.